+++ frontend/src/Pages/Admin/index.js
@@ -1,0 +1,1 @@
+export * from './Admin'
