+++ frontend/src/Pages/Cart/PaymentPage.js
@@ -1,0 +1,66 @@
+import React, { useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { useLocation, useNavigate } from "react-router-dom";
+import { savePaymentMethod } from "~/actions/cartActions";
+import CheckoutSteps from "~/components/CheckoutSteps";
+
+export default function PaymentPage() {
+    var props = useLocation();
+    const navigate = useNavigate();
+
+    const cart = useSelector((state) => state.cart);
+    const { shippingAddress } = cart;
+    if (!shippingAddress.address) {
+        navigate("/shipping");
+    }
+    const [paymentMethod, setPayment] = useState("PayPal");
+    const dispatch = useDispatch();
+    const submitHandler = (e) => {
+        e.preventDefault();
+        dispatch(savePaymentMethod(paymentMethod));
+        navigate("/placeorder");
+    };
+    return (
+        <div>
+            <CheckoutSteps step1 step2 step3></CheckoutSteps>
+            <form className="form" onSubmit={submitHandler}>
+                <div>
+                    <h1>Payment Method</h1>
+                </div>
+                <div>
+                    <div>
+                        <input
+                            type="radio"
+                            id="paypal"
+                            value="PayPal"
+                            name="paymentMethod"
+                            required
+                            checked
+                            onChange={(e) => setPayment(e.target.value)}
+                        ></input>
+                        <label htmlFor="paypal">PayPal</label>
+                    </div>
+                </div>
+                <div>
+                    <div>
+                        <input
+                            type="radio"
+                            id="stripe"
+                            value="Stripe"
+                            name="paymentMethod"
+                            required
+                            onChange={(e) => setPayment(e.target.value)}
+                        ></input>
+                        <label htmlFor="stripe">Stripe</label>
+                    </div>
+                </div>
+                <div>
+                    <label />
+                    <button className="primary" type="submit">
+                        Continue
+                    </button>
+                </div>
+            </form>
+        </div>
+    );
+}
