+++ frontend/src/Pages/Cart/index.js
@@ -1,1 +1,1 @@
-export { default } from "./CartPage";
+export * from "./CartPage";
