--- frontend/src/Pages/Dashboard/index.js
@@ -1,1 +1,0 @@
-export { default } from './DashboardPage'
