+++ frontend/src/Pages/Order/index.js
@@ -1,1 +1,1 @@
-export { default } from './OrderPage'
+export * from "./Order";
