--- frontend/src/Pages/OrderHistory/index.js
@@ -1,1 +1,0 @@
-export { default } from './OrderHistoryPage'
