--- frontend/src/Pages/OrderList/index.js
@@ -1,1 +1,0 @@
-export { default } from './OrderListPage'
