--- frontend/src/Pages/Payment/index.js
@@ -1,1 +1,0 @@
-export { default } from './PaymentPage'
