--- frontend/src/Pages/PlaceOrder/index.js
@@ -1,1 +1,0 @@
-export { default } from './PlaceOrderPage'
