+++ frontend/src/Pages/Product/index.js
@@ -1,1 +1,1 @@
-export { default } from './ProductPage'
+export * from './Product'
