--- frontend/src/Pages/ProductEdit/index.js
@@ -1,1 +1,0 @@
-export { default } from './ProductEditPage'
