--- frontend/src/Pages/ProductList/ProductListPage.js
@@ -1,133 +1,0 @@
-import React, { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { createProduct, deleteProduct, listProducts } from "~/actions/productActions";
-import LoadingBox from "~/components/LoadingBox";
-import MessageBox from "~/components/MessageBox";
-import { Link, useLocation, useNavigate, useParams } from "react-router-dom";
-import { PRODUCT_CREATE_RESET, PRODUCT_DELETE_RESET } from "~/constants/productConstants";
-
-export default function ProductListPage() {
-    const userSignin = useSelector((state) => state.userSignin);
-    const { pageNumber = 1 } = useParams();
-    const { userInfo } = userSignin;
-    const location = useLocation();
-    const sellerMode = location.search === "?seller";
-    const navigate = useNavigate();
-    const dispatch = useDispatch();
-    const productList = useSelector((state) => state.productList);
-    const productCreate = useSelector((state) => state.productCreate);
-    const { loading, error, products, page, pages } = productList;
-    console.log("productList", productList);
-    const {
-        loading: loadingCreate,
-        error: errorCreate,
-        success: successCreate,
-        product: createdProduct,
-    } = productCreate;
-
-    const productDelete = useSelector((state) => state.productDelete);
-    const { loading: loadingDelete, error: errorDelete, success: successDelete } = productDelete;
-
-    console.log("productCreate", productCreate);
-    useEffect(() => {
-        if (successCreate) {
-            dispatch({ type: PRODUCT_CREATE_RESET });
-            navigate(`/product/${createdProduct._id}/edit`);
-        }
-        if (successDelete) {
-            dispatch({ type: PRODUCT_DELETE_RESET });
-        }
-        dispatch(listProducts({ seller: sellerMode ? userInfo._id : "", pageNumber }));
-    }, [
-        createdProduct,
-        dispatch,
-        navigate,
-        sellerMode,
-        successCreate,
-        successDelete,
-        userInfo._id,
-        pageNumber,
-    ]);
-    const deleteHandler = (product) => {
-        /// TODO: dispatch delete action
-        if (window.confirm("Are you sure to delete?")) {
-            dispatch(deleteProduct(product._id));
-        }
-    };
-    const createHandler = () => {
-        /// TODO dispatch add acttion
-        dispatch(createProduct());
-    };
-
-    return (
-        <div>
-            <div className="row">
-                <button type="button" className="primary" onClick={createHandler}>
-                    Create Product
-                </button>
-            </div>
-            {loadingCreate && <LoadingBox></LoadingBox>}
-            {errorCreate && <MessageBox variant="danger">{errorCreate}</MessageBox>}
-            {loading ? (
-                <LoadingBox></LoadingBox>
-            ) : error ? (
-                <MessageBox variant="danger">{error}</MessageBox>
-            ) : (
-                <>
-                    <table className="table">
-                        <thead>
-                            <tr>
-                                <th>ID</th>
-                                <th>NAME</th>
-                                <th>PRICE</th>
-                                <th>CATEGORY</th>
-                                <th>BRAND</th>
-                                <th>ACTIONS</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {products.map((product) => (
-                                <tr key={product._id}>
-                                    <td>{product._id}</td>
-                                    <td>{product.name}</td>
-                                    <td>{product.price}</td>
-                                    <td>{product.category}</td>
-                                    <td>{product.brand}</td>
-                                    <td>
-                                        <button
-                                            type="button"
-                                            className="small"
-                                            onClick={() =>
-                                                navigate(`/product/${product._id}/edit`)
-                                            }
-                                        >
-                                            Edit
-                                        </button>
-                                        <button
-                                            type="button"
-                                            className="small"
-                                            onClick={() => deleteHandler(product)}
-                                        >
-                                            Delete
-                                        </button>
-                                    </td>
-                                </tr>
-                            ))}
-                        </tbody>
-                    </table>
-                    <div className="row center pagination">
-                        {[...Array(pages).keys()].map((x) => (
-                            <Link
-                                className={x + 1 === page ? "active" : ""}
-                                key={x + 1}
-                                to={`/productlist/pageNumber/${x + 1}`}
-                            >
-                                {x + 1}
-                            </Link>
-                        ))}
-                    </div>
-                </>
-            )}
-        </div>
-    );
-}
