--- frontend/src/Pages/ProductList/index.js
@@ -1,1 +1,0 @@
-export { default } from './ProductListPage'
