--- frontend/src/Pages/Register/index.js
@@ -1,1 +1,0 @@
-export { default } from './RegisterPage'
