--- frontend/src/Pages/SearchPage/index.js
@@ -1,1 +1,0 @@
-export { default } from './SearchPage'
