--- frontend/src/Pages/Shipping/index.js
@@ -1,1 +1,0 @@
-export { default } from './ShippingAddressPage'
