--- frontend/src/Pages/Support/index.js
@@ -1,1 +1,0 @@
-export { default } from "./SupportPage";
