+++ frontend/src/Pages/User/index.js
@@ -1,0 +1,1 @@
+export * from "./User";
