--- frontend/src/Pages/UserEdit/index.js
@@ -1,1 +1,0 @@
-export { default } from './UserEditPage'
