--- frontend/src/Pages/UserList/index.js
@@ -1,1 +1,0 @@
-export { default } from './UserList'
